Start removal of tool usage of SkSurface::MakeFromBackendTextureAsRenderTarget.
Adds a new helper that creates a GrBackendRenderTarget using
GrGpu and then wraps it in a SkSurface. Uses the SkSurface
release proc to delete the BERT using GrGpu.

Upgrades GrGpu::createTestingOnlyBackendRenderTarget to create MSAA
buffers and to clear to an initial color.

Updates many tests/tool to call sites to use the helper instead of
SkSurface::MakeFromBackendTextureAsRenderTarget.

+++ tests/SurfaceTest.cpp
@@ -30,6 +30,7 @@
 #include "src/image/SkSurface_Gpu.h"
 #include "tests/Test.h"
 #include "tests/TestUtils.h"
+#include "tools/gpu/BackendSurfaceFactory.h"
 
 #include <functional>
 #include <initializer_list>
@@ -116,12 +117,6 @@ DEF_GPUTEST_FOR_RENDERING_CONTEXTS(GrContext_colorTypeSupportedAsSurface, report
             REPORTER_ASSERT(reporter, can == SkToBool(surf), "ct: %d, can: %d, surf: %d",
                             colorType, can, SkToBool(surf));
 
-            surf = SkSurface::MakeFromBackendTextureAsRenderTarget(context, backendTex,
-                                                                   kTopLeft_GrSurfaceOrigin, 1,
-                                                                   colorType, nullptr, nullptr);
-            REPORTER_ASSERT(reporter, can == SkToBool(surf), "ct: %d, can: %d, surf: %d",
-                            colorType, can, SkToBool(surf));
-
             surf.reset();
             context->flushAndSubmit();
             context->deleteBackendTexture(backendTex);
@@ -157,44 +152,27 @@ DEF_GPUTEST_FOR_RENDERING_CONTEXTS(GrContext_colorTypeSupportedAsSurface, report
                                 "Should store an allowed sample count (%d vs %d)", allowedCnt,
                                 storedCnt);
             }
-
-            surf = SkSurface::MakeFromBackendTextureAsRenderTarget(context, backendTex,
-                                                                   kTopLeft_GrSurfaceOrigin,
-                                                                   kSampleCnt, colorType,
-                                                                   nullptr, nullptr);
-            REPORTER_ASSERT(reporter, can == SkToBool(surf),
-                            "colorTypeSupportedAsSurface:%d, surf:%d, ct:%d", can, SkToBool(surf),
-                            colorType);
-            if (surf) {
-                auto rtc = ((SkSurface_Gpu*)(surf.get()))->getDevice()->accessRenderTargetContext();
-                int storedCnt = rtc->numSamples();
-                int allowedCnt = context->priv().caps()->getRenderTargetSampleCount(
-                        storedCnt, backendTex.getBackendFormat());
-                REPORTER_ASSERT(reporter, storedCnt == allowedCnt,
-                                "Should store an allowed sample count (%d vs %d)", allowedCnt,
-                                storedCnt);
-            }
-
             surf.reset();
             context->flushAndSubmit();
             context->deleteBackendTexture(backendTex);
         }
 
-        {
-            auto* gpu = context->priv().getGpu();
-
-            GrBackendRenderTarget backendRenderTarget = gpu->createTestingOnlyBackendRenderTarget(
-                    16, 16, SkColorTypeToGrColorType(colorType));
-            bool can = context->colorTypeSupportedAsSurface(colorType);
-            auto surf = SkSurface::MakeFromBackendRenderTarget(context, backendRenderTarget,
-                                                               kTopLeft_GrSurfaceOrigin, colorType,
-                                                               nullptr, nullptr);
-            REPORTER_ASSERT(reporter, can == SkToBool(surf), "ct: %d, can: %d, surf: %d", colorType,
-                            can, SkToBool(surf));
-            surf.reset();
-            context->flushAndSubmit();
-            if (backendRenderTarget.isValid()) {
-                gpu->deleteTestingOnlyBackendRenderTarget(backendRenderTarget);
+        for (int sampleCnt : {1, 2}) {
+            auto surf = MakeBackendRenderTargetSurface(context, {16, 16}, sampleCnt,
+                                                       kTopLeft_GrSurfaceOrigin, colorType);
+            bool can = context->colorTypeSupportedAsSurface(colorType) &&
+                       context->maxSurfaceSampleCountForColorType(colorType) >= sampleCnt;
+            REPORTER_ASSERT(reporter, can == SkToBool(surf), "ct: %d, sc: %d, can: %d, surf: %d",
+                            colorType, sampleCnt, can, SkToBool(surf));
+            if (surf) {
+                auto rtc = ((SkSurface_Gpu*)(surf.get()))->getDevice()->accessRenderTargetContext();
+                auto backendFormat = rtc->asSurfaceProxy()->backendFormat();
+                int storedCnt = rtc->numSamples();
+                int allowedCnt = context->priv().caps()->getRenderTargetSampleCount(storedCnt,
+                                                                                    backendFormat);
+                REPORTER_ASSERT(reporter, storedCnt == allowedCnt,
+                                "Should store an allowed sample count (%d vs %d)", allowedCnt,
+                                storedCnt);
             }
         }
     }
@@ -738,28 +716,22 @@ static bool supports_readpixels(const GrCaps* caps, SkSurface* surface) {
     return caps->surfaceSupportsReadPixels(rt) == GrCaps::SurfaceReadPixelsSupport::kSupported;
 }
 
-static sk_sp<SkSurface> create_gpu_surface_backend_texture_as_render_target(
-    GrContext* ctx, int sampleCnt, const SkColor4f& color, GrBackendTexture* outTexture) {
-
+static sk_sp<SkSurface> create_gpu_surface_backend_render_target(GrContext* ctx,
+                                                                 int sampleCnt,
+                                                                 const SkColor4f& color,
+                                                                 GrBackendTexture* outTexture) {
     const int kWidth = 10;
     const int kHeight = 10;
 
-    SkImageInfo ii = SkImageInfo::Make(kWidth, kHeight, SkColorType::kRGBA_8888_SkColorType,
-                                       kPremul_SkAlphaType);
-
-    if (!CreateBackendTexture(ctx, outTexture, ii, color, GrMipMapped::kNo, GrRenderable::kYes)) {
+    SkPMColor4f pmColor{color.fR, color.fG, color.fB, color.fA};
+    auto surf = MakeBackendRenderTargetSurface(ctx, {kWidth, kHeight}, sampleCnt,
+                                               kTopLeft_GrSurfaceOrigin, kRGBA_8888_SkColorType,
+                                               nullptr, nullptr, &pmColor);
+    if (!surf) {
         return nullptr;
     }
-
-    sk_sp<SkSurface> surface = SkSurface::MakeFromBackendTextureAsRenderTarget(
-            ctx, *outTexture, kTopLeft_GrSurfaceOrigin, sampleCnt, kRGBA_8888_SkColorType,
-            nullptr, nullptr);
-
-    if (!surface) {
-        DeleteBackendTexture(ctx, *outTexture);
-        return nullptr;
-    }
-    return surface;
+    surf->getCanvas()->clear(color);
+    return surf;
 }
 
 static void test_surface_context_clear(skiatest::Reporter* reporter,
@@ -824,8 +796,8 @@ DEF_GPUTEST_FOR_GL_RENDERING_CONTEXTS(SurfaceClear_Gpu, reporter, ctxInfo) {
 
     // Wrapped RTs are *not* supposed to clear (to allow client to partially update a surface).
     const SkColor4f kOrigColor{.67f, .67f, .67f, 1};
-    for (auto& surfaceFunc : {&create_gpu_surface_backend_texture,
-                              &create_gpu_surface_backend_texture_as_render_target}) {
+    for (auto& surfaceFunc :
+         {&create_gpu_surface_backend_texture, &create_gpu_surface_backend_render_target}) {
         GrBackendTexture backendTex;
         auto surface = surfaceFunc(context, 1, kOrigColor, &backendTex);
         if (!surface) {
@@ -901,8 +873,8 @@ DEF_GPUTEST_FOR_RENDERING_CONTEXTS(SurfacePartialDraw_Gpu, reporter, ctxInfo) {
 
     static const SkColor4f kOrigColor { 0.667f, 0.733f, 0.8f, 1 };
 
-    for (auto& surfaceFunc : {&create_gpu_surface_backend_texture,
-                              &create_gpu_surface_backend_texture_as_render_target}) {
+    for (auto& surfaceFunc :
+         {&create_gpu_surface_backend_texture, &create_gpu_surface_backend_render_target}) {
         // Validate that we can draw to the canvas and that the original texture color is
         // preserved in pixels that aren't rendered to via the surface.
         // This works only for non-multisampled case.
@@ -958,7 +930,7 @@ DEF_GPUTEST_FOR_RENDERING_CONTEXTS(SurfaceWrappedWithRelease_Gpu, reporter, ctxI
                                                         ReleaseChecker::Release,
                                                         &releaseChecker);
         } else {
-            backendRT = gpu->createTestingOnlyBackendRenderTarget(kWidth, kHeight,
+            backendRT = gpu->createTestingOnlyBackendRenderTarget({kWidth, kHeight},
                                                                   GrColorType::kRGBA_8888);
             if (!backendRT.isValid()) {
                 continue;
@@ -1005,8 +977,8 @@ DEF_GPUTEST_FOR_GL_RENDERING_CONTEXTS(SurfaceAttachStencil_Gpu, reporter, ctxInf
 
     auto resourceProvider = context->priv().resourceProvider();
 
-    for (auto& surfaceFunc : {&create_gpu_surface_backend_texture,
-                              &create_gpu_surface_backend_texture_as_render_target}) {
+    for (auto& surfaceFunc :
+         {&create_gpu_surface_backend_texture, &create_gpu_surface_backend_render_target}) {
         for (int sampleCnt : {1, 4, 8}) {
             GrBackendTexture backendTex;
             auto surface = surfaceFunc(context, sampleCnt, kOrigColor, &backendTex);
@@ -1021,7 +993,9 @@ DEF_GPUTEST_FOR_GL_RENDERING_CONTEXTS(SurfaceAttachStencil_Gpu, reporter, ctxInf
             GrRenderTarget* rt = surface->getCanvas()
                 ->internal_private_accessTopLayerRenderTargetContext()->accessRenderTarget();
             REPORTER_ASSERT(reporter, resourceProvider->attachStencilAttachment(rt, sampleCnt));
-            context->deleteBackendTexture(backendTex);
+            if (backendTex.isValid()) {
+                context->deleteBackendTexture(backendTex);
+            }
         }
     }
 }
